requirements.txt
anyio==4.7.0
certifi==2024.12.14
convertdate==2.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
persian==0.5.0
PyMeeus==0.5.12
python-dotenv==1.0.1
python-telegram-bot==21.9
sniffio==1.3.1
